last fixes

+++ src/companymanager.h
@@ -11,11 +11,11 @@
 
 
 // companies
-struct company_1 { char name[23]; int payment;}; // Hederlige Harrys Bilar
-struct company_2 { char name[22]; int payment;}; // Farmor Ankas Pajer AB
-struct company_3 { char name[27]; int payment;}; // Svarte Petters Svartbyggen
-struct company_4 { char name[24]; int payment;}; // Långbens detektivbyrå
-struct IOT_Commercial { char name[18]; int payment;}; // IOT
+struct company_1 { char name[23]; double payment;}; // Hederlige Harrys Bilar
+struct company_2 { char name[22]; double payment;}; // Farmor Ankas Pajer AB
+struct company_3 { char name[27]; double payment;}; // Svarte Petters Svartbyggen
+struct company_4 { char name[24]; double payment;}; // Långbens detektivbyrå
+struct IOT_Commercial { char name[18]; double payment;}; // IOT
 
 int paymentSumCalc();
 void freqcalc(int FQ[], int P);
